backend/requirements.txt
Flask==2.1.0
Werkzeug<3.0.0
gunicorn==20.1.0
flatlib==0.2.3
pyahocorasick==1.4.4